requirements.txt
Authlib==0.15.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
idna==3.3
pycparser==2.21
python-dotenv==0.19.2
requests==2.26.0
urllib3==1.26.7
